fix the issue to negative numbers

+++ db/migrate/20221003173306_create_foods.rb
@@ -6,7 +6,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
       t.string :name
       t.string :measure_unit
       t.float :price, default: 0
-      t.integer :quantity, default: 0
+      t.float :quantity, default: 0
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
